1. Separated CsvWriter and CsvReader 2. changed restore snapshots to restore-snapshots
