+++ src/main/java/seedu/address/logic/parser/CliSyntax.java
@@ -41,7 +41,7 @@ public class CliSyntax {
     public static final String COMMAND_IMPORT = "import";
     public static final String COMMAND_EXPORT = "export";
     public static final String COMMAND_MAIL = "mail";
-    public static final String COMMAND_SNAPSHOTS = "snapshots";
+    public static final String COMMAND_SNAPSHOTS = "restore-snapshots";
     public static final String COMMAND_SCHEDULE = "schedule";
     public static final String COMMAND_SCHEDULE_ADD = "schedule-add";
     public static final String COMMAND_SCHEDULE_DELETE = "schedule-delete";
