+++ src/main/java/seedu/address/logic/parser/RestoreCommandParser.java
@@ -2,13 +2,12 @@
 package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.COMMAND_SNAPSHOTS;
 
 import java.io.IOException;
 
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.BackUpCommand;
 import seedu.address.logic.commands.RestoreCommand;
-import seedu.address.logic.commands.RestoreSnapshotsCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.backup.BackupList;
 
@@ -25,13 +24,9 @@ public class RestoreCommandParser implements Parser<RestoreCommand> {
     public RestoreCommand parse(String args) throws ParseException {
         args = args.trim();
         try {
-            BackupList backupList = ParserUtil.parseBackup(args);
-            if (args.equals(COMMAND_SNAPSHOTS)) {
-                return new RestoreSnapshotsCommand(backupList);
-            } else {
-                Index index = ParserUtil.parseIndex(args);
-                return new RestoreCommand(backupList, index);
-            }
+            BackupList backupList = new BackupList(BackUpCommand.DEST_PATH);
+            Index index = ParserUtil.parseIndex(args);
+            return new RestoreCommand(backupList, index);
         } catch (ParseException pe) {
             throw new ParseException(
                     String.format(MESSAGE_INVALID_COMMAND_FORMAT, RestoreCommand.MESSAGE_USAGE), pe);
